refactor: deployment service mock
