+++ src/app/shared/services/crashlytics/crashlytics.service.spec.ts
@@ -2,14 +2,14 @@ import { TestBed } from "@angular/core/testing";
 
 import { CrashlyticsService } from "./crashlytics.service";
 import { DeploymentService } from "../deployment/deployment.service";
-import { MockDeploymentService } from "../deployment/deployment.service.spec";
+import { MockDeploymentService } from "../deployment/deployment.service.mock.spec";
 
 describe("CrashlyticsService", () => {
   let service: CrashlyticsService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: [{ provide: DeploymentService, useValue: new MockDeploymentService({}) }],
+      providers: [{ provide: DeploymentService, useValue: new MockDeploymentService() }],
     });
     service = TestBed.inject(CrashlyticsService);
   });
