+++ src/app/shared/services/error-handler/error-handler.service.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from "@angular/core/testing";
 import { ErrorHandlerService } from "./error-handler.service";
 import { FirebaseService } from "../firebase/firebase.service";
 import { DeploymentService } from "../deployment/deployment.service";
-import { MockDeploymentService } from "../deployment/deployment.service.spec";
+import { MockDeploymentService } from "../deployment/deployment.service.mock.spec";
 
 describe("ErrorHandlerService", () => {
   let service: ErrorHandlerService;
@@ -15,7 +15,7 @@ describe("ErrorHandlerService", () => {
           provide: FirebaseService,
           useValue: {},
         },
-        { provide: DeploymentService, useValue: new MockDeploymentService({}) },
+        { provide: DeploymentService, useValue: new MockDeploymentService() },
       ],
     });
     service = TestBed.inject(ErrorHandlerService);
